add Frontend changes for Star rating

+++ backend/src/main/java/com/github/esgoet/backend/dto/BookDto.java
@@ -13,7 +13,7 @@ public record BookDto(
         String description,
         String isbn,
         String cover,
-        double rating,
+        int rating,
         LocalDate publicationDate
 ) {
 }
