+++ backend/src/main/java/com/github/esgoet/backend/models/Book.java
@@ -15,7 +15,7 @@ public record Book(
         String description,
         String isbn,
         String cover,
-        double rating,
+        int rating,
         LocalDate publicationDate
 ) {
 }
